% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_score.R
\name{get_mi_score}
\alias{get_mi_score}
\title{get MI score for a given study}
\usage{
get_mi_score(
  studyid = NULL,
  path_db = NULL,
  fake_study = FALSE,
  master_CompileData = NULL,
  score_in_list_format = FALSE,
  xpt_dir = NULL
)
}
\arguments{
\item{studyid}{Optional when \code{xpt_dir} is set; required for SQLite. Study identifier.}

\item{path_db}{Optional when \code{xpt_dir} is set; path of database.}

\item{fake_study}{optional, Boolean \cr
whether study generated by SENDsanitizer package}

\item{master_CompileData}{optional; precomputed compile data.}

\item{score_in_list_format}{optional; return format.}

\item{xpt_dir}{Optional; path to a directory containing XPT files for one study (flat: \code{xpt_dir/mi.xpt}, etc.).}
}
\value{
score
}
\description{
Get microscopic findings score for a study. Data from SQLite or from a directory of XPT files (flat layout).
}
\examples{
\dontrun{
get_mi_score(studyid = '1234123', path_db = 'path/to/database.db')
get_mi_score(xpt_dir = '/path/to/study_folder')
}
}
